Move shared types to root module
